% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Compiegne_env}
\alias{Compiegne_env}
\title{dataset used for session 4}
\format{
An object of class \code{data.frame} with 26 rows and 8 columns.
}
\usage{
data(Compiegne_env)
}
\description{
dataset used for session 4
}
\keyword{datasets}
