% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dune_spec}
\alias{dune_spec}
\title{dataset used for session 4}
\format{
An object of class \code{data.frame} with 3 rows and 31 columns.
}
\usage{
data(dune_spec)
}
\description{
dataset used for session 4
}
\keyword{datasets}
