% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GutterTor_Ellenberg}
\alias{GutterTor_Ellenberg}
\title{dataset used for assignment 4}
\format{
An object of class \code{data.frame} with 24 rows and 8 columns.

An object of class \code{data.frame} with 24 rows and 8 columns.
}
\usage{
data(GutterTor_Ellenberg)

data(GutterTor_Ellenberg)
}
\description{
dataset used for assignment 4

dataset used for assgnment 4
}
\keyword{datasets}
