% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GutterTor_env}
\alias{GutterTor_env}
\title{dataset used for assignment 4}
\format{
An object of class \code{data.frame} with 25 rows and 6 columns.

An object of class \code{data.frame} with 25 rows and 6 columns.
}
\usage{
data(GutterTor_env)

data(GutterTor_env)
}
\description{
dataset used for assignment 4

dataset used for assgnment 4
}
\keyword{datasets}
