% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_s5}
\alias{iris_s5}
\title{dataset used for session 5}
\format{
An object of class \code{data.frame} with 30 rows and 5 columns.
}
\usage{
data(iris_s5)
}
\description{
dataset used for session 5
}
\keyword{datasets}
