% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kembel_sp}
\alias{kembel_sp}
\title{dataset used for session 5}
\format{
An object of class \code{data.frame} with 27 rows and 77 columns.
}
\usage{
data(kembel_sp)
}
\description{
dataset used for session 5
}
\keyword{datasets}
