% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leuven_a3}
\alias{leuven_a3}
\title{dataset used for assignment 3}
\format{
An object of class \code{data.frame} with 69 rows and 238 columns.
}
\usage{
data(leuven_a3)
}
\description{
dataset used for assignment 3
}
\keyword{datasets}
