% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meerdaal}
\alias{meerdaal}
\title{dataset used for session 5}
\format{
An object of class \code{data.frame} with 13 rows and 3 columns.
}
\usage{
data(meerdaal)
}
\description{
dataset used for session 5
}
\keyword{datasets}
