% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nphk_a3}
\alias{nphk_a3}
\title{dataset used for assignment 3}
\format{
An object of class \code{data.frame} with 18 rows and 40 columns.
}
\usage{
data(nphk_a3)
}
\description{
dataset used for assignment 3
}
\keyword{datasets}
