% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant_growth}
\alias{plant_growth}
\title{dataset used for session 5}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
data(plant_growth)
}
\description{
dataset used for session 5
}
\keyword{datasets}
