% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varechem_s2}
\alias{varechem_s2}
\title{dataset used for session 2}
\format{
An object of class \code{data.frame} with 24 rows and 10 columns.
}
\usage{
data(varechem_s2)
}
\description{
dataset used for session 2
}
\keyword{datasets}
