% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varespec_s6}
\alias{varespec_s6}
\title{dataset used for session 6}
\format{
An object of class \code{data.frame} with 24 rows and 45 columns.
}
\usage{
data(varespec_s6)
}
\description{
dataset used for session 6
}
\keyword{datasets}
