% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegdata1}
\alias{vegdata1}
\title{First dataset to be used in assignment 1}
\format{
An object of class \code{data.frame} with 72 rows and 16 columns.
}
\usage{
vegdata1
}
\description{
First dataset to be used in assignment 1
}
\keyword{datasets}
