% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegdata2}
\alias{vegdata2}
\title{Second dataset to be used in assignment 1}
\format{
An object of class \code{data.frame} with 72 rows and 16 columns.
}
\usage{
vegdata2
}
\description{
Second dataset to be used in assignment 1
}
\keyword{datasets}
