% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegdata_a2}
\alias{vegdata_a2}
\title{dataset used for assignment 2}
\format{
An object of class \code{data.frame} with 144 rows and 17 columns.
}
\usage{
data(vegdata_a2)
}
\description{
dataset used for assignment 2
}
\keyword{datasets}
